+++ packages/extension-base/src/services/balance-service/transfer/xcm/utils.ts
@@ -3,7 +3,13 @@
 
 import { _ChainAsset, _ChainInfo } from '@subwallet/chain-list/types';
 import { fetchParaSpellChainMap } from '@subwallet/extension-base/constants/paraspell-chain-map';
+import { _isSnowBridgeXcm } from '@subwallet/extension-base/core/substrate/xcm-parser';
+import { _isAcrossChainBridge } from '@subwallet/extension-base/services/balance-service/transfer/xcm/acrossBridge';
+import { isAvailChainBridge } from '@subwallet/extension-base/services/balance-service/transfer/xcm/availBridge';
 import { CreateXcmExtrinsicProps } from '@subwallet/extension-base/services/balance-service/transfer/xcm/index';
+import { _isPolygonChainBridge } from '@subwallet/extension-base/services/balance-service/transfer/xcm/polygonBridge';
+import { _isPosChainBridge } from '@subwallet/extension-base/services/balance-service/transfer/xcm/posBridge';
+import { _isPureEvmChain } from '@subwallet/extension-base/services/chain-service/utils';
 import { ProxyServiceRoute } from '@subwallet/extension-base/types/environment';
 import { fetchFromProxyService } from '@subwallet/extension-base/utils';
 import BigNumber from 'bignumber.js';
@@ -38,7 +44,7 @@ export type DryRunResult = {
   hops: THopInfo[]
 }
 
-interface GetXcmFeeRequest {
+export interface GetXcmFeeRequest {
   sender: string,
   recipient: string,
   value: string,
@@ -78,7 +84,9 @@ const paraSpellApi = {
   buildXcm: `${version}/x-transfer`,
   feeXcm: `${version}/xcm-fee`,
   dryRunXcm: `${version}/dry-run`,
-  dryRunPreviewXcm: `${version}/dry-run-preview`
+  dryRunPreviewXcm: `${version}/dry-run-preview`,
+  maxTransferable: `${version}/transferable-amount`,
+  minTransferable: `${version}/min-transferable-amount`
 };
 
 function txHexToSubmittableExtrinsic (api: ApiPromise, hex: string): SubmittableExtrinsic<'promise'> {
@@ -335,6 +343,42 @@ export async function estimateXcmFee (request: GetXcmFeeRequest) {
   return await response.json() as GetXcmFeeResult;
 }
 
+export async function fetchMinXcmTransferableAmount (request: GetXcmFeeRequest) {
+  const { fromChainInfo: originChain, fromTokenInfo: originTokenInfo, recipient, sender, toChainInfo: destinationChain, value: sendingValue } = request;
+  const paraSpellChainMap = await fetchParaSpellChainMap();
+  const paraSpellIdentifyV4 = originTokenInfo.metadata?.paraSpellIdentifyV4;
+
+  if (!paraSpellIdentifyV4) {
+    throw new Error('Token is not support XCM at this time');
+  }
+
+  const bodyData = {
+    senderAddress: sender,
+    address: recipient,
+    from: paraSpellChainMap[originChain.slug],
+    to: paraSpellChainMap[destinationChain.slug],
+    currency: createParaSpellCurrency(paraSpellIdentifyV4, sendingValue),
+    options: {
+      abstractDecimals: false
+    }
+  };
+
+  const response = await fetchFromProxyService(
+    ProxyServiceRoute.PARASPELL,
+    paraSpellApi.minTransferable,
+    {
+      method: 'POST',
+      body: JSON.stringify(bodyData),
+      headers: {
+        'Content-Type': 'application/json',
+        Accept: 'application/json'
+      }
+    }
+  );
+
+  return await response.json() as string;
+}
+
 function createParaSpellCurrency (paraSpellIdentifyV4: Record<string, any>, amount: string): ParaSpellCurrency {
   return {
     ...paraSpellIdentifyV4,
@@ -352,4 +396,42 @@ export function isChainNotSupportDryRun (str: string): boolean {
   const regex = /(?=.*DryRunApi)(?=.*not available).*/i; // Example: DryRunApi is not available on node Acala
 
   return regex.test(str);
+}
+
+export function isSubstrateCrossChain (originChainInfo: _ChainInfo, destinationChainInfo: _ChainInfo) {
+  if (originChainInfo.slug === destinationChainInfo.slug) {
+    return false;
+  }
+
+  // isAvailBridgeFromEvm
+  if (_isPureEvmChain(originChainInfo) && isAvailChainBridge(destinationChainInfo.slug)) {
+    return false;
+  }
+
+  // isAvailBridgeFromAvail
+  if (isAvailChainBridge(originChainInfo.slug) && _isPureEvmChain(destinationChainInfo)) {
+    return false;
+  }
+
+  // isSnowBridgeEvmTransfer
+  if (_isPureEvmChain(originChainInfo) && _isSnowBridgeXcm(originChainInfo, destinationChainInfo)) {
+    return false;
+  }
+
+  // isPolygonBridgeTransfer
+  if (_isPolygonChainBridge(originChainInfo.slug, destinationChainInfo.slug)) {
+    return false;
+  }
+
+  // isPosBridgeTransfer
+  if (_isPosChainBridge(originChainInfo.slug, destinationChainInfo.slug)) {
+    return false;
+  }
+
+  // isAcrossBridgeTransfer
+  if (_isAcrossChainBridge(originChainInfo.slug, destinationChainInfo.slug)) {
+    return false;
+  }
+
+  return true;
 }