+++ packages/extension-koni-ui/src/Popup/Confirmations/variants/Transaction/variants/TransferBlock.tsx
@@ -3,12 +3,16 @@
 
 import { ExtrinsicDataTypeMap, ExtrinsicType } from '@subwallet/extension-base/background/KoniTypes';
 import { _isAcrossChainBridge } from '@subwallet/extension-base/services/balance-service/transfer/xcm/acrossBridge';
-import { _getChainName } from '@subwallet/extension-base/services/chain-service/utils';
+import { isSubstrateCrossChain } from '@subwallet/extension-base/services/balance-service/transfer/xcm/utils';
+import { _getAssetDecimals, _getAssetPriceId, _getAssetSymbol, _getChainName } from '@subwallet/extension-base/services/chain-service/utils';
 import { AlertBox, CommonTransactionInfo } from '@subwallet/extension-koni-ui/components';
 import MetaInfo from '@subwallet/extension-koni-ui/components/MetaInfo/MetaInfo';
 import QuoteRateDisplay from '@subwallet/extension-koni-ui/components/Swap/QuoteRateDisplay';
+import { BN_TEN } from '@subwallet/extension-koni-ui/constants';
 import { useGetAccountByAddress, useGetNativeTokenBasicInfo } from '@subwallet/extension-koni-ui/hooks';
 import { RootState } from '@subwallet/extension-koni-ui/stores';
+import { Number } from '@subwallet/react-ui';
+import BigN from 'bignumber.js';
 import CN from 'classnames';
 import React, { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
@@ -23,6 +27,8 @@ const Component: React.FC<Props> = ({ className, transaction }: Props) => {
   const { t } = useTranslation();
   const data = transaction.data as ExtrinsicDataTypeMap[ExtrinsicType.TRANSFER_BALANCE];
   const xcmData = transaction.data as ExtrinsicDataTypeMap[ExtrinsicType.TRANSFER_XCM];
+  const priceMap = useSelector((root: RootState) => root.price.priceMap);
+  const { currencyData } = useSelector((root: RootState) => root.price);
   const chainInfoMap = useSelector((root: RootState) => root.chainStore.chainInfoMap);
   const assetRegistryMap = useSelector((root: RootState) => root.assetRegistry.assetRegistry);
   const tokenInfo = assetRegistryMap[transaction.extrinsicType === ExtrinsicType.TRANSFER_XCM ? xcmData.tokenSlug : data.tokenSlug];
@@ -49,6 +55,86 @@ const Component: React.FC<Props> = ({ className, transaction }: Props) => {
   const chainInfo = useMemo(() => chainInfoMap[transaction.chain], [chainInfoMap, transaction.chain]);
   const { decimals: nativeTokenDecimals, symbol: nativeTokenSymbol } = useGetNativeTokenBasicInfo(transaction.chain);
   const feeInfo = transaction.estimateFee;
+  const crossChainFeeInfo = transaction?.xcmDestinationFee;
+
+  const priceNativeValue = useMemo(() => {
+    const nativeTokenSlug = `${transaction.chain}-NATIVE-${nativeTokenSymbol}`;
+    const nativeTokenInfo = assetRegistryMap[nativeTokenSlug];
+
+    const priceNativeId = _getAssetPriceId(nativeTokenInfo);
+
+    return priceMap[priceNativeId] || 0;
+  }, [assetRegistryMap, nativeTokenSymbol, priceMap, transaction.chain]);
+
+  const transferTokenSymbol = _getAssetSymbol(tokenInfo);
+  const transferTokenDecimals = _getAssetDecimals(tokenInfo);
+
+  const transferTokenValue = useMemo(() => {
+    const transferTokenPriceId = _getAssetPriceId(tokenInfo);
+
+    return priceMap[transferTokenPriceId] || 0;
+  }, [priceMap, tokenInfo]);
+
+  const destTransferTokenValue = useMemo(() => {
+    const destTransferTokenPriceId = _getAssetPriceId(destTokenInfo);
+
+    return priceMap[destTransferTokenPriceId] || 0;
+  }, [destTokenInfo, priceMap]);
+
+  const convertedFeeValueToUSD = useMemo(() => {
+    if (!feeInfo?.value) {
+      return 0;
+    }
+
+    return new BigN(feeInfo?.value)
+      .multipliedBy(priceNativeValue)
+      .dividedBy(BN_TEN.pow(nativeTokenDecimals || 0))
+      .toNumber();
+  }, [feeInfo, nativeTokenDecimals, priceNativeValue]);
+
+  const convertedCrossChainFeeValueToUSD = useMemo(() => {
+    if (!crossChainFeeInfo?.value) {
+      return 0;
+    }
+
+    return new BigN(crossChainFeeInfo?.value)
+      .multipliedBy(transferTokenValue)
+      .dividedBy(BN_TEN.pow(tokenInfo.decimals || 0))
+      .toNumber();
+  }, [crossChainFeeInfo, tokenInfo, transferTokenValue]);
+
+  const convertedAmountValueToUSD = useMemo(() => {
+    if (!data.value) {
+      return 0;
+    }
+
+    return new BigN(data.value)
+      .multipliedBy(transferTokenValue)
+      .dividedBy(BN_TEN.pow(tokenInfo.decimals || 0))
+      .toNumber();
+  }, [data.value, tokenInfo.decimals, transferTokenValue]);
+
+  const convertedXcmDestTokenValueToUSD = useMemo(() => {
+    if (!isAcrossBridge || !xcmData.metadata?.amountOut) {
+      return 0;
+    }
+
+    return new BigN(xcmData.metadata?.amountOut)
+      .multipliedBy(destTransferTokenValue)
+      .dividedBy(BN_TEN.pow(destTokenInfo.decimals || 0))
+      .toNumber();
+  }, [destTokenInfo.decimals, destTransferTokenValue, isAcrossBridge, xcmData.metadata?.amountOut]);
+
+  const isSubstrateXcm = useMemo(() => {
+    const originChainInfo = chainInfoMap[originChainSlug];
+    const destinationChainInfo = chainInfoMap[destinationChainSlug];
+
+    if (!originChainInfo || !destinationChainInfo) {
+      return false;
+    }
+
+    return isSubstrateCrossChain(originChainInfo, destinationChainInfo);
+  }, [destinationChainSlug, originChainSlug, chainInfoMap]);
 
   return (
     <>
@@ -100,7 +186,11 @@ const Component: React.FC<Props> = ({ className, transaction }: Props) => {
             ))
       }
 
-      <MetaInfo hasBackgroundWrapper>
+      <MetaInfo
+        hasBackgroundWrapper
+        labelColorScheme={'gray'}
+        valueColorScheme={'light'}
+      >
         {!!transaction.wrappingStatus &&
           <MetaInfo.Account
             address={data.to}
@@ -117,32 +207,83 @@ const Component: React.FC<Props> = ({ className, transaction }: Props) => {
               <QuoteRateDisplay
                 className={'__quote-estimate-swap-value'}
                 fromAssetInfo={tokenInfo}
-                rateValue={Number(xcmData.metadata.rate)}
+                rateValue={parseFloat(xcmData.metadata.rate)}
                 toAssetInfo={destTokenInfo}
               />
             </MetaInfo.Default>
-            <MetaInfo.Number
-              decimals={destTokenInfo.decimals || 0}
-              label={t('ui.TRANSACTION.Confirmations.TransferBlock.expectedAmount')}
-              suffix={destTokenInfo.symbol}
-              value={xcmData.metadata.amountOut}
-            />
+            <MetaInfo.Default label={t('ui.TRANSACTION.Confirmations.TransferBlock.expectedAmount')}>
+              <Number
+                decimal={destTokenInfo.decimals || 0}
+                suffix={destTokenInfo.symbol}
+                value={xcmData.metadata.amountOut}
+              />
+              <Number
+                className={'__amount-price-value'}
+                decimal={0}
+                prefix={`~ ${(currencyData.isPrefix && currencyData.symbol) || ''}`}
+                suffix={(!currencyData.isPrefix && currencyData.symbol) || ''}
+                value={convertedXcmDestTokenValueToUSD}
+              />
+            </MetaInfo.Default>
           </>
           : (
-            <MetaInfo.Number
-              decimals={tokenInfo.decimals || 0}
+            <MetaInfo.Default
               label={t('ui.TRANSACTION.Confirmations.TransferBlock.amount')}
-              suffix={tokenInfo.symbol}
-              value={data.value || 0}
-            />
+            >
+              <Number
+                decimal={tokenInfo.decimals || 0}
+                suffix={tokenInfo.symbol}
+                value={data.value || 0}
+              />
+              <Number
+                className={'__amount-price-value'}
+                decimal={0}
+                prefix={`~ ${(currencyData.isPrefix && currencyData.symbol) || ''}`}
+                suffix={(!currencyData.isPrefix && currencyData.symbol) || ''}
+                value={convertedAmountValueToUSD}
+              />
+            </MetaInfo.Default>
           )}
 
-        {!transaction.wrappingStatus && <MetaInfo.Number
-          decimals={feeInfo ? feeInfo.decimals : nativeTokenDecimals}
+        <MetaInfo.Default
           label={t('ui.TRANSACTION.Confirmations.TransferBlock.networkFee')}
-          suffix={feeInfo ? feeInfo.symbol : nativeTokenSymbol}
-          value={feeInfo ? feeInfo.value : 0}
-        />}
+        >
+          {!transaction.wrappingStatus && <div className='__fee-editor-area'>
+            <Number
+              decimal={feeInfo ? feeInfo.decimals : nativeTokenDecimals}
+              suffix={feeInfo ? feeInfo.symbol : nativeTokenSymbol}
+              value={feeInfo ? feeInfo.value : 0}
+            />
+
+            <Number
+              className={'__fee-price-value'}
+              decimal={0}
+              prefix={`~ ${(currencyData.isPrefix && currencyData.symbol) || ''}`}
+              suffix={(!currencyData.isPrefix && currencyData.symbol) || ''}
+              value={convertedFeeValueToUSD}
+            />
+          </div>}
+        </MetaInfo.Default>
+
+        {transaction.extrinsicType === ExtrinsicType.TRANSFER_XCM && crossChainFeeInfo?.value && isSubstrateXcm && (
+          <MetaInfo.Default label={t('ui.TRANSACTION.Confirmations.TransferBlock.crossChainFee')}>
+            <div className={'__value-col-wrapper'}>
+              <Number
+                decimal={transferTokenDecimals}
+                suffix={transferTokenSymbol}
+                value={crossChainFeeInfo ? crossChainFeeInfo?.value : 0}
+              />
+
+              <Number
+                className={'__fee-price-value'}
+                decimal={0}
+                prefix={`~ ${(currencyData.isPrefix && currencyData.symbol) || ''}`}
+                suffix={(!currencyData.isPrefix && currencyData.symbol) || ''}
+                value={convertedCrossChainFeeValueToUSD}
+              />
+            </div>
+          </MetaInfo.Default>
+        )}
       </MetaInfo>
       {
         transaction.extrinsicType === ExtrinsicType.TRANSFER_XCM &&
@@ -161,6 +302,11 @@ const Component: React.FC<Props> = ({ className, transaction }: Props) => {
 
 export const TransferBlock = styled(Component)<Props>(({ theme: { token } }: Props) => {
   return {
+    '.__fee-price-value, .__amount-price-value': {
+      fontSize: `${token.fontSizeSM}px !important`,
+      lineHeight: '20px !important',
+      color: `${token.colorTextTertiary} !important`
+    },
     '&.alert-area': {
       marginTop: token.marginSM
     }